Reformat Cache module
